don't allow deletion/holding of locked jobs
Reviewed-on: https://gerrit.instructure.com/62435
Tested-by: Jenkins
QA-Review: August Thornton <august@instructure.com>

+++ lib/delayed/backend/redis/bulk_update.lua
@@ -21,20 +21,30 @@ else
   end
 end
 
+local count = 0
 for idx, job_id in ipairs(ids) do
   if action == 'hold' then
-    local queue, strand = unpack(redis.call('HMGET', Keys.job(job_id), 'queue', 'strand'))
-    remove_from_queues(job_id, queue, strand)
-    redis.call('HMSET', Keys.job(job_id), 'locked_at', now, 'locked_by', 'on hold', 'attempts', 50)
+    local queue, strand, locked_by = unpack(redis.call('HMGET', Keys.job(job_id), 'queue', 'strand', 'locked_by'))
+    if not locked_by then
+      count = count + 1
+      remove_from_queues(job_id, queue, strand)
+      redis.call('HMSET', Keys.job(job_id), 'locked_at', now, 'locked_by', 'on hold', 'attempts', 50)
+    end
   elseif action == 'unhold' then
     local queue, locked_by = unpack(redis.call('HMGET', Keys.job(job_id), 'queue', 'locked_by'))
-    add_to_queues(job_id, queue, now)
-    redis.call('HDEL', Keys.job(job_id), 'locked_at', 'locked_by')
-    redis.call('HMSET', Keys.job(job_id), 'attempts', 0)
+    if locked_by == 'on hold' then
+      count = count + 1
+      add_to_queues(job_id, queue, now)
+      redis.call('HDEL', Keys.job(job_id), 'locked_at', 'locked_by')
+      redis.call('HMSET', Keys.job(job_id), 'attempts', 0)
+    end
   elseif action == 'destroy' then
-    destroy_job(job_id, now)
+    local locked_by = redis.call('HGET', Keys.job(job_id), 'locked_by')
+    if not locked_by or locked_by == 'on hold' then
+      count = count + 1
+      destroy_job(job_id, now)
+    end
   end
 end
 
--- returns the # of jobs matching the query, not necessarily the # whose state was changed
-return table.getn(ids)
+return count
